refactor: extract header subcomponents
